主要コードレビュー対応
レビュー対応中

レビューの内主要な物に対応OK

+++ Editor/DynamicTile/TileModule/TileCleanupTempFolder.cs
@@ -38,6 +38,11 @@ namespace PLATEAU.Editor.DynamicTile.TileModule
         private void CleanupTempFolder()
         {
             var assetPath = DynamicTileProcessingContext.PrefabsTempSavePath;
+            if (!AssetDatabase.IsValidFolder(assetPath))
+            {
+                Debug.Log($"一時フォルダーなし: {assetPath}"); // Assets内のケース
+                return;
+            }
             if (AssetDatabase.DeleteAsset(assetPath))
             {
                 AssetDatabase.Refresh();
@@ -45,7 +50,7 @@ namespace PLATEAU.Editor.DynamicTile.TileModule
             }
             else
             {
-                Debug.Log($"一時フォルダーなし: {assetPath}"); // Assets内のケース
+                Debug.LogWarning("一時フォルダの削除に失敗しました: " + assetPath);
             }
         }
     }
